Splits Utility Service
This commit splits the utility server into seperate files to  make future development easier
