+++ src/utilities/index.js
@@ -1,0 +1,17 @@
+'use strict';
+
+const {
+  fieldData,
+  recordId,
+  namespace,
+  filterResponse,
+  sanitizeParameters
+} = require('filemaker.utilities');
+
+module.exports = {
+  fieldData,
+  recordId,
+  namespace,
+  filterResponse,
+  sanitizeParameters
+};
